Finish OP::FEE operation in rest of operations layer, use 64 bit for fees in contract

+++ src/TAO/Operation/contract.cpp
@@ -95,7 +95,7 @@ namespace TAO
 
 
         /*  Get the fees for contract. */
-        const int32_t& Contract::Fees() const
+        const int64_t& Contract::Fees() const
         {
             return nFees;
         }
